Rename time package to nettime to avoid stdlib conflict
- Rename agent/time/ directory to agent/nettime/
- Change package declaration from 'time' to 'nettime' in time_sync.go
- Update all imports to use 'nettime "github.com/netwatcherio/netwatcher-agent/nettime"'
- Change time.AdjustedTime() calls to nettime.AdjustedTime()
- Remove alias needed - nettime package used directly without alias
