+++ workers/data.go
@@ -5,8 +5,8 @@ import (
 	"encoding/json"
 	"time"
 
+	nettime "github.com/netwatcherio/netwatcher-agent/nettime"
 	"github.com/netwatcherio/netwatcher-agent/probes"
-	"github.com/netwatcherio/netwatcher-agent/time"
 	"github.com/netwatcherio/netwatcher-agent/web"
 	log "github.com/sirupsen/logrus"
 )
@@ -28,7 +28,7 @@ func ProbeDataWorker(wsH *web.WSClient, ch chan probes.ProbeData) {
 
 	go func(cn *web.WSClient, c chan probes.ProbeData) {
 		for p := range ch {
-			p.CreatedAt = time.AdjustedTime()
+			p.CreatedAt = nettime.AdjustedTime()
 			marshal, err := json.Marshal(p)
 			if err != nil {
 				log.Errorf("ProbeDataWorker: failed to marshal probe data: %v", err)
